+++ src/backend/github/issue-comment/issue-comment.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
-import { createWebhookApp } from "../../src/common/test-utils/probot";
+import { createWebhookApp } from "../../../common/test-utils/probot";
 
 describe("GitHub Actions", () => {
 	let app;
@@ -8,7 +8,7 @@ describe("GitHub Actions", () => {
 	describe("issue_comment", () => {
 		describe("created", () => {
 			it("should update the GitHub issue with a linked Jira ticket", async () => {
-				const fixture = require("../fixtures/issue-comment-basic.json");
+				const fixture = require("../../../../test/fixtures/issue-comment-basic.json");
 
 				jiraNock
 					.get("/rest/api/latest/issue/TEST-123?fields=summary")
